+++ src/dipdup/config/__init__.py
@@ -46,7 +46,6 @@ from dipdup.exceptions import ConfigInitializationException
 from dipdup.exceptions import ConfigurationError
 from dipdup.exceptions import FrameworkException
 from dipdup.exceptions import IndexAlreadyExistsError
-from dipdup.models import LoggingValues
 from dipdup.models import ReindexingAction
 from dipdup.models import ReindexingReason
 from dipdup.models import SkipHistory
@@ -584,6 +583,9 @@ class AdvancedConfig:
     metrics: MetricsLevel = MetricsLevel.basic
     alt_operation_matcher: bool = False
 
+    class Config:
+        extra = 'allow'
+
 
 @dataclass
 class DipDupConfig:
@@ -622,7 +624,7 @@ class DipDupConfig:
     prometheus: PrometheusConfig | None = None
     advanced: AdvancedConfig = field(default_factory=AdvancedConfig)
     custom: dict[str, Any] = field(default_factory=dict)
-    logging: LoggingValues = LoggingValues.default
+    logging: dict[str, str | int] | str | int = 'INFO'
 
     def __post_init_post_parse__(self) -> None:
         if self.package != pascal_to_snake(self.package):
@@ -748,13 +750,23 @@ class DipDupConfig:
         return datasource
 
     def set_up_logging(self) -> None:
-        level = {
-            LoggingValues.default: logging.INFO,
-            LoggingValues.quiet: logging.WARNING,
-            LoggingValues.verbose: logging.DEBUG,
-        }[self.logging]
-        logging.getLogger('dipdup').setLevel(level)
-        logging.getLogger(self.package).setLevel(level)
+        loglevels = self.logging
+        if not isinstance(loglevels, dict):
+            loglevels = {
+                'dipdup': loglevels,
+                self.package: loglevels,
+            }
+
+        for name, level in loglevels.items():
+            try:
+                if isinstance(level, str):
+                    level = getattr(logging, level.upper())
+                if not isinstance(level, int):
+                    raise ValueError
+            except (AttributeError, ValueError):
+                raise ConfigurationError(f'Invalid logging level `{level}` for logger `{name}`') from None
+
+            logging.getLogger(name).setLevel(level)
 
     def initialize(self) -> None:
         self._set_names()