+++ src/dipdup/config/tezos_tzkt_token_transfers.py
@@ -28,8 +28,8 @@ class TzktTokenTransfersHandlerConfig(HandlerConfig):
 
     contract: TezosContractConfig | None = None
     token_id: int | None = None
-    # FIXME: Can't use `from_` field alias in dataclass; fixed in dipdup.yaml instead
-    # FIXME: See https://github.com/pydantic/pydantic/issues/4286
+    # FIXME: Can't use `from_` field alias in dataclass
+    # FIXME: See https://github.com/pydantic/pydantic/issues/4286 (fixed in upcoming v2)
     from_: TezosContractConfig | None = None
     to: TezosContractConfig | None = None